:truck: Move shadcn generated utils to ./utils
