+++ apps/next/src/components/ui/dialog.tsx
@@ -4,7 +4,7 @@ import * as React from "react"
 import * as DialogPrimitive from "@radix-ui/react-dialog"
 import { X } from "lucide-react"
 
-import { cn } from "@/lib/utils"
+import { cn } from "@/utils/tw"
 
 const Dialog = DialogPrimitive.Root
