+++ apps/next/src/components/ui/skeleton.tsx
@@ -1,4 +1,4 @@
-import { cn } from "@/lib/utils"
+import { cn } from "@/utils/tw"
 
 function Skeleton({
   className,
